# 277 - speed up build/run by setting up outputs for gradle UI build